Added PUT and DELETE methods

+++ src/main/java/com/RESTful/CRUD/CustomerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-interface CustomerRepository extends JpaRepository<Customer, Long> {
+interface CustomerRepository extends JpaRepository<Customer, Integer> {
     
 }
